Authorize.Net: Remove supported countries except for AU CA US
There was a request that we remove all countries except United States,
Canada, and Australia from the Authorize.net gateway guide.

https://docs.spreedly.com/payment-gateways/authorize-net/
This is due to PSD2/SCA in Europe. Authorize.net will not add
any new customers for the countries being removed due to this change.

This PR will only affect documentation.

CE-368