BPoint: Remove amount from void requests
The `void` method for the BPoint gateway was previously implemented
incorrectly. Throughout Active Merchant, gateway `void` requests are
passed only an authorization number for the original transaction and an
options hash. However, for BPoint, the `void` method was also passed an
explicit `amount`. Due to the implementations on top of Active Merchant,
this was causing issues for Spreedly customers.

The BPoint `void` method has been rewritten to take in only
`authorization` and an `options` hash. Because the BPoint gateway
requires the amount to be sent on void transactions, this value should
be sent within the `options` hash with the key `amount`.

CE-338

Unit:
19 tests, 63 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote:
18 tests, 41 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

All unit tests:
4439 tests, 71411 assertions, 0 failures, 0 errors, 0 pendings, 2
omissions, 0 notifications
100% passed