Worldpay: Remove unnecessary .tag! methods
When building an XML request, the `.tag!` method is only required if the
XML tag is a non-standard string, e.g., if it contains a hyphen or a
colon. Therefore, this method is often not needed and can be replaced
with a far more concise version.

This commit removes all unnecessary `.tag!` method calls for the
Worldpay gateway to improve the readability of the code.

Unit:
69 tests, 411 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote:
56 tests, 240 assertions, 2 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
96.4286% passed

The two remote test failures are previously failing tests:
 - `test_3ds_version_1_parameters_pass_thru`
 - `test_3ds_version_2_parameters_pass_thru`

All unit tests:
4439 tests, 71411 assertions, 0 failures, 0 errors, 0 pendings, 2
omissions, 0 notifications
100% passed