Merchant Warrior: Send void amount in options
The `void` method for the Merchant Warrior gateway was previously
implemented incorrectly. Throughout Active Merchant, gateway `void`
requests are passed only an authorization number for the original
transaction and an options hash. However, for Merchant Warrior, the void
method was also passed an explicit amount. Due to the implementations on
top of Active Merchant, this was causing issues for Spreedly customers.

The Merchant Warrior `void` method has been rewritten to take in only
authorization and an options hash. Because the Merchant Warrior gateway
requires the amount to be sent on void transactions, this value should
be sent within the `options` hash with the key `amount`.

In addition, the required verification hash for `void` transactions
should consist of the original transaction ID, not the amount and the
currency. This change also adds a new `void_verification_hash` method
that calculates the correct hash for the `void` method.

CE-379

Unit:
17 tests, 92 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote:
15 tests, 81 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

All unit tests:
4442 tests, 71468 assertions, 0 failures, 0 errors, 0 pendings, 2
omissions, 0 notifications
100% passed