RuboCop: Fix Layout/IndentHash
This fixes the RuboCop to-do for the indentation of the first key inside
of a hash literal where the brace and the first key are on separate
lines. To better align with the existing code conventions, this cop was
set to use the `consistent` style instead of the default
`special_inside_parentheses` style.

All unit tests:
4444 tests, 71479 assertions, 0 failures, 0 errors, 0 pendings, 2
omissions, 0 notifications
100% passed