Bambora Apac: Send void amount in options
The `void` method for the Bambora Apac gateway was previously
implemented incorrectly. Throughout Active Merchant, gateway `void`
requests are passed only an authorization number for the original
transaction and an options hash. However, for Bambora Apac, the void
method was also passed an explicit amount. This break in the established
pattern was causing issues for implementations on top of Active
Merchant.

The Bambora Apac `void` method has been rewritten to take in only
authorization and an options hash. Because the Bambora Apac gateway
requires the amount to be sent on void transactions, this value should
be sent within the `options` hash with the key `amount`.

CE-389

Unit:
7 tests, 48 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0
notifications
100% passed

Remote:
15 tests, 30 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

All unit tests:
4444 tests, 71480 assertions, 0 failures, 0 errors, 0 pendings, 2
omissions, 0 notifications
100% passed