Stripe: Add connected account support
CE-425

Adds support for the following fields on the Stripe gateway for Connect
transactions:
- transfer_data[amount, destination]
- on_behalf_of
- transfer_group
- application_fee_amount

This was done by moving the `add_connected_account` function from the
Stripe Payment Intents gateway to the Stripe gateway which it inherits
from. One change had to be made to the implementation: since the Stripe
Payment Intents gateway called the `application_fee_amount` field
`application_fee`, which is already a field in the Stripe gateway, both
gateways are to call this field `application_fee_amount` moving forward
with the Stripe Payment Intents gateway having code in place to also
allow it to be called `application_fee` so that there are no breaking
changes.