Stripe Payment Intents: Updates for idempotency key support
Stripe Payment Intents supports the use of an idempotency key, but
that key can only be used against a single endpoint. Because the
Stripe PI flow can touch multiple different endpoints, we need to
ensure that it is slightly different across the different types
of requests to prevent errors. For instance, if the payment method
provided is a credit card, the adapter hits the payment method endpoint
to create that object before hitting the payment intent endpoint to
create the intent. This PR adds a method to add a suffix aligning to
request endpoints other than /payment_intents (e.g. a post to /payment_methods
would be <provided-key-pm> or a post to /payment_intents/intent-id/capture
would be <provided-key-capture>) to ensure that keys are unique but stable
across a request type. Also adds a remote test to ensure idempotency key
behavior remains as expected when a credit card payment type is used.

Unit:
9 tests, 55 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
34 tests, 158 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed