Orbital: Fix stored credentials
ECS-1076

https://secure.paymentech.com/developercenter/files/file?fid=J4p8Vagjmx4%3D
Subsequent cardholder initiated stored credential transactions are failing
due to the `get_msg_type` method looking for the `customer` parameter.

This change corrects the method to look for the `cardholder` parameter in
addition to the `customer` parameter in order to return the correct `MITMsgType`
required by the gateway and to keep backwards compatibility.

Unit:
91 tests, 527 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
34 tests, 188 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

All unit tests:
4459 tests, 71574 assertions, 0 failures, 0 errors, 0 pendings, 2 omissions, 0 notifications
100% passed