Iridium: Localize zero-decimal currencies
ECS-1119

Iridium (PayVector) requires the transaction amount to be sent in "minor
currency". This change ensures that zero-decimal currencies are properly
sent (i.e. 14200 JPY is sent as 142)

A new remote test was not written because Iridium (PayVector) does not
return the amount in the response.

Unit:
13 tests, 56 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
21 tests, 84 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

All unit tests:
4462 tests, 71592 assertions, 0 failures, 0 errors, 0 pendings, 2 omissions, 0 notifications
100% passed