Add merchant descriptor methods from main AM (#102)
Why:
Summary for requirement:

Merchants must add a descriptor that indicates a transaction related to a trial offer to the first transaction processed after the trial offer ends. This descriptor should be added to the Merchant Name field of the Clearing Record and should include language like “trial,” “trial period,” “free trial,” and the like. This will then appear on bank statements, banking apps, and text alerts that the cardholder sees.

What is the enhanced descriptor for? Where is it required? (Visa)

The enhanced descriptor (e.g., “trial,” “trial period,” “free trial”) is to be included in the Merchant Name field of the clearing record for the first transaction at the end of a trial period. It is not required for subsequent transactions. This descriptor will then appear on cardholder statements, online banking, mobile apps and SMS / text alerts, in the same way discretionary data or additional invoice / order numbers appear for e-commerce transactions today, to identify the nature of the transaction. Visa is not restricting the word choice of the enhanced descriptor, as long as the merchant can identify that it is a trial-related transaction for the cardholder and issuer.

Additional info from visa

Statement Descriptor An additional descriptor indicating a trial period-related transaction will be required in the Merchant Name field of the Clearing Record for the first transaction at the end of a trial period. This descriptor (for example, “trial,” “trial period,” “free trial”) will then appear on cardholder statements, online banking, mobile apps, and SMS/text alerts, in the same way discretionary, additional invoice/order numbers appear for electronic commerce transactions.

What:
Required technical changes (from cybersource documentation)

You can provide a merchant descriptor to be displayed on the customer’s bank account statement. The descriptor includes your company’s name and a description of the product or service that was purchased.
The merchant descriptor field overrides the corresponding value in your CyberSource account. If you do not include this field in the request, CyberSource uses the company name from your merchant account.
Before sending a merchant descriptor with a payment or credit request, check with your processor to find out if you need to register your merchant descriptor information with them.
The merchantInformation.merchantDescriptor.name field requires a particular format:
Characters 1-15: name of your company. If the name is fewer than 15 characters, use spaces to fill in the full 15 characters. If the name is more than 15 characters, provide only the first 15 characters of the name.
Characters 16-25: description of the product or service.

Please note - when I used the above instruction, the payment did not process, coming back with an error of incorrectly formatted merchant_descriptor field. I used simple trial ended because of that.