CheckoutV2: Send appropriate error messages for 4xx errors
When we're unable to parse an error message from the returned body
of a checkout response, we send the generic message `Unable to read
error message`. This causes confusion, and in a lot of cases there
is more information we could be providing to help inform users of the
problem. Specifically, when there's a 4xx error, we attempt to parse
the response body, but can ultimately get a ParserError which will end
up setting the AM message to a generic error message. We've seen with
401 errors, the body is simply an empty string, so this default happens
everytime and customers have to dig to figure out exactly what happened.
This PR updates the logic to instead set an error type and code when
rescuing the ParserError (if available). That leads us to present a more
informative error in such cases (e.g. `401: Unavailable`). It also updates
the default condition in `message_from` to check for the value from the
response['message'] field if present before setting the message to `Unable
to read...`, which notes that there's an issue with the JSON (which also
seems more informative).

Remote:
33 tests, 80 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Unit:
29 tests, 130 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed