[forte] Use underscore for unused argument in test
Working with the Forte gateway in a fork, I received this error from Codacy:

```
test/unit/gateways/forte_test.rb
Unused block argument - `headers`. If it's necessary, use `_` or `_headers` as an argument name to indicate that it won't be used.
end.check_request do |type, url, parameters, headers|
```

I assume the rule was added since this file was last modified.

This will mark the `headers` argument as unused so that the quality review will pass.