HPS: Prevent NoMethodError when account_type or account_holder_type nil
For check payment methods, the HPS adapter attempts to upcase the
account_type and account_holder_type fields when adding them to
the request. If either of these is missing, a NoMethodError is raised.
This PR makes an update to safely call upcase in these instances
to prevent such errors.

Remote (same failures on master):
50 tests, 125 assertions, 5 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
90% passed

Unit:
54 tests, 267 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed