Set a default collectionIndicator
For mastercard cards, we need to always be passing a collectionIndicator for the request to succeed.

I decided to use `2` as the default collectionIndicator, as it was already the one used for tokenized card and
it seems to match our usecase from all the pther choices:

- 0: UCAF collection is not supported at your web
site.
- 1: UCAF collection is supported at your web
site, and the UCAF was populated.
- 2: UCAF collection is supported at your web
site and the UCAF was populated. This value
indicates a successful Mastercard Identity
Check transaction.
- 5: UCAF collection is supported at your web
site, and the UCAF was populated based on the
risk assessment that the issuer performed. This
value is supported only for Masterpass
transactions.
- 6: UCAF collection is supported at your web
site, and the UCAF was populated based on the
risk assessment that you performed. This value
is supported only for Masterpass transactions.