[Cybersource] Send cavv as xid when xid is missing
We've been experiencing an error with 3DS 2.0 where Cybersource still
requires an xid even though this field isn't present for 3DS 2.0 transactions.

This is the exact error `<c:missingField>c:xid</c:missingField>`

After discussing with Cybersource and Cardinal it was decided that we
will send the cavv as the xid for 3DS 2.0 until they sort out their api.