AAFES: Integrating purchase method for the AAFES gateway
First iteration of integrating AAFES gateway into AM starting with the
basic purchase method.

Some nuance for how this gateway works means there is no auth method
(unless otherwise approved by AAFES - an arm of the DoD). What this
means for anyone who wants to use this gateway is that they must become
an approved partner with AAFES and implement the Authentication portion
on their own internal systems.

Additionally, AAFES is both the card issuer and the gateway provider for
the MilStar card. The MilStar card is a credit card only available to
members of the US armed forces. A part of the authentication step, is a
vaulting process, where a MilStar card holder and merchant customer is
redirected from the merchant checkout to the AAFES Stargate (AAFES
gateway vault) and will redirect a tokenized version of the credit card
along with an expiration date (YYMM) and zipcode for verification in
subsequent transactions with the gateway.

On the topic of transactions. AAFES identifies each transaction by an
RRN value that is required to be 12 characters long and base64 encoded.
This value MUST be unique for each transaction and has a life cycle of
30 days on the AAFFES gateway. Capture, Refund, and Void transactions
all require a unique RRN for the transaction as well as an original RRN
to identify the initial Purchase transaction.

It is also worth noting, this gateway requires the currency value to be
passed as a base 10 decimal value as opposed to an integer.
CE-636

Unit: 2 tests, 8 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote: 2 tests, 4 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed