Element: Fix unit tests
A couple of unit tests for the Element gateway were previously
implemented as remote tests, which I discovered last night when the
Element gateway was down and the unit test file failed.

To ensure that the unit test file is not dependent on the gateway being
up, these two tests using remote calls to the gateway have been replaced
with actual unit tests that stub out the communication with the gateway.

Unit:
19 tests, 72 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote:
21 tests, 55 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

All unit tests:
4520 tests, 72098 assertions, 0 failures, 0 errors, 0 pendings, 0
omissions, 0 notifications
100% passed