Elavon: Upgrade gateway integration [WIP]
This is the start of the work necessary to upgrade the Elavon gateway
integration to move from the deprecated `process.do` endpoint to the
more current XML endpoint `processxml.do`.

This WIP commit includes the major updates to the gateway structure to
use XML and to achieve a successful remote purchase request.

Many of the tests still need to be updated, including updating all of
the stubbed responses in the unit tests.

CE-704

`ruby -Itest test/remote/gateways/remote_elavon_test.rb -n
test_successful_purchase`

1 tests, 5 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0
notifications
100% passed