FirstdataE4V27Gateway: Fix strip_line_breaks method
The recently added strip_line_breaks method added the requirement that
all values in the shipping hash be strings. It is often the case though
that nil values are supplied, especially for fields like address2. This
fix changes the function to allow for nil values.

Modified the following test to introduce a nil value to make sure this
does not regress: test_requests_scrub_newline_and_return_characters_from
_verification_string_components

4533 tests, 72185 assertions, 0 failures, 0 errors, 0 pendings,
0 omissions, 0 notifications
100% passed
692 files inspected, no offenses detected