Cybersource: Ensure Partner Solution Id placement conforms to schema
The Cybersource XML request must follow the defined xsd schema format
or else it will result in an XML parse error. Currently, if someone
tries to use stored credential fields in addition to the partner
solution id, they will encounter an error. The schema requires
`subsequentAuth` to come before the solution id and for the other stored
credential fields to come after solution id. This PR updates the request
building methods to ensure that these fields are included in the right
order, also adding remote tests for this behavior.

Remote:
89 tests, 457 assertions, 5 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
94.382% passed
(pre-existing failures)

Unit:
96 tests, 463 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed