Quickbooks: Omit empty strings in address
This fixes a bug caused by submitting an empty string for the state
field (which is submitted as 'region' to Quickbooks). While `nil` values
for this field are fine because the field is then simply omitted from
the request sent to Quickbooks, sending an empty string results in the
error message `card.address.region is invalid`.

ECS-1355

Unit:
22 tests, 118 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote:
18 tests, 44 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

All unit tests:
4544 tests, 72251 assertions, 0 failures, 0 errors, 0 pendings, 0
omissions, 0 notifications
100% passed