HPS: Enable refunds using capture txn
Currently, if you try to issue a refund using a capture as your
reference transaction, you will get an error. This is because the
authorization value here only refers to the capture, with no
connection to the auth. In other adapters, we solve for this by
building up an authorization value that includes references to the
current transaction as well as reference transactions (and potentially
other data). This allows us to parse this field to return the
correct reference transaction for supplemental actions. This PR
updates HPS to build an authorization value including the authorize
transaction id for captures. It also adds in logic to pluck off
the appropriate value when issuing a refund on a capture txn.

Unit:
54 tests, 267 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
52 tests, 143 assertions, 3 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
94.2308% passed
(preexisting failures also on master)