CyberSource: Always send default address
Previously, when an empty string was submitted for the
`[:billing_address][:country]`, CyberSource would response with an error
message since this field is required.

When a payment method that contained only a phone number with no other
address information was used, this resulted in the payment failing. To
resolve this issue, the `setup_address_hash` method was updated to match
the intent of the already existing comment: to always supply the default
address to CyberSource regardless of whether a `nil` value or an empty
string was initially submitted as the value for `country` or any other
required address field.

ECS-1363

Unit:
97 tests, 470 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote:
95 tests, 490 assertions, 5 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
94.7368% passed

All unit tests:
4551 tests, 72299 assertions, 0 failures, 0 errors, 0 pendings, 0
omissions, 0 notifications
100% passed