WIP: Adds e-check option for Orbital
This is resurrects an old branch from 2017 attempting to add `echeck` capability to the Orbital gateway. It is a work-in-progress and is not polished. There are no unit tests. The effort was abandoned when the requesting customer decided to move on. I was not able to get the one remote test passing although, at the time, the response indicted the test data was not correct.

Remote Tests:

There were 3 failures: the expected `echeck` failure plus 2 3DS failures for American Express that I did not troubleshoot at all. Not sure that they would be impacted by these additions but I did not investigate.

38 tests, 193 assertions, 3 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
92.1053% passed