Orbital: Fix schema errors
After fixing an issue with problematic values being sent for network
tokenized PMs, errors relating to the request not adhering to the
DTD would sometimes appear. This happens in the existing code when
soft descriptors are passed along with a network tokenized card.
In the schema, the soft descriptor fields need to come before the
network token fields. This PR updates the order of these elements
and adds additional test conditions.

Unit:
96 tests, 563 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
37 tests, 192 assertions, 2 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
94.5946% passed
(same as master)