Orbital: Resolve CardIndicators issue
CardIndicators field is being sent, but not received by the Orbital gateway. In the schema, the line items field needs to come before the card indicators field, but they are switched. This PR updates the order of those elements and adds additional testing.

CE-710

UNIT:
96 tests, 563 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed
REMOTE:
38 tests, 195 assertions, 4 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
89.4737% passed
(Same as master)

Update CHANGELOG