CyberSource: Handle IP address field in tokenized purchase requests (#133)
While https://github.com/activemerchant/active_merchant/pull/1875 introduces the `ipAddress` field, the whole method is only used for requests that send full payment profile data and not its tokenized form.

According to the documentation, it should be possible to pass billing address data to the tokenized purchase request, though.

While it'd be ideal to introduce the ability to do that, it'd require a bigger refactoring. The current logic treats lack of some params as if the params should not be added while lack of some other params as if they should be added with a nil value. In case of the tokenized purchase, the passed billing address should be more of a "patch" than a "put" so that we don't require the whole address to be passed each time, but rather only the fields that need to be updated.

That's why I ended up implementing the `ipAddress` part of the billing address only.