Braintree: Add support for $0 auth verification
If a merchant sends the field `allow_card_verification` as `true`, the `verify` method will utilize Braintree's credit-card-verification endpoint. If `allow_card_verification` is sent as `false` or not sent at all, it will execute the `verify` with the existing authorize/void transaction flow as it has in the past. To note: `allow_card_verification` is not a gateway supported field so it is not passed on to Braintree during the `verify` process. The purpose of the field is to direct which flow the `verify` method should use.

CE-555

Local:
4697 tests, 73368 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed
Unit:
82 tests, 188 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed
Remote:
83 tests, 442 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed