[DigitalRiver] Add DigitalRiver Gateway - store and purchase (#152)
This PR adds support for DigitalRiver - store and purchase actions.

It uses the digital_river gem which is a wrapper around DigitalRiver API created by Chargify.

The store action creates a Customer resource if one does not already exist and attaches a Source to that Customer.

Th purchase action checks the order state and returns a failure if the order is not in accepted state. If it is, it creates a fulfillment and returns the capture id for the order. For now we assume that there will only be one charge per one order.

The remote tests cannot test the first of the purchase paths due to the lack of a test card that would provide such a scenario.
The unit test covers it instead.